+++ tests/__init__.py
@@ -2,5 +2,5 @@ import os
 import sys
 
 PROJECT_PATH = os.getcwd()
-SOURCE_PATH = os.path.join(PROJECT_PATH, "gladiator")
+SOURCE_PATH = os.path.join(PROJECT_PATH, "jericho")
 sys.path.append(SOURCE_PATH)